Pack GlyphMetric to 12 bytes, matching the SD font-pack metric layout
The usize offset/len fields padded every glyph metric to 16 bytes; the
SD font-pack format (FONT_PACK_METRIC_BYTES) already proves
offset: u32, len: u16 suffice. Adopting that layout in-flash shrinks
the 49,802 built-in metric entries by 199,208 B of .rodata on both
boards (3.11 -> 2.91 MB), the generated tables recompile unchanged, and
decode_metric becomes a near-identity copy. Golden frames and reading
goldens pass byte-identical on X4 and X3.

+++ fw/src/custom_font.rs
@@ -387,7 +387,7 @@ where
         for y in 0..metric.height as usize {
             let offset = face
                 .bitmap_offset
-                .checked_add(metric.offset as u32)?
+                .checked_add(metric.offset)?
                 .checked_add((y * row_bytes) as u32)?;
             self.file.seek_from_start(offset).ok()?;
             read_full(self.file, &mut row[..row_bytes])?;
@@ -479,8 +479,8 @@ where
 
 fn decode_metric(bytes: &[u8; FONT_PACK_METRIC_BYTES]) -> GlyphMetric {
     GlyphMetric {
-        offset: u32::from_le_bytes([bytes[0], bytes[1], bytes[2], bytes[3]]) as usize,
-        len: u16::from_le_bytes([bytes[4], bytes[5]]) as usize,
+        offset: u32::from_le_bytes([bytes[0], bytes[1], bytes[2], bytes[3]]),
+        len: u16::from_le_bytes([bytes[4], bytes[5]]),
         width: bytes[6],
         height: bytes[7],
         x_offset: bytes[8] as i8,
